requirements.txt
google-cloud-storage
pandas
plotly
Pillow